+++ src/timers.js
@@ -13,9 +13,13 @@ p.t.interval = function(f, ms) {
 
 p.t.dt = function() {
     p.t.t2 = da();
-    p.t.rdt = 1000 / (p.t.t2 - p.t.t1);
-    p.t.t1 = p.t.t2;
+    p.t.rdt = p.t.t2 - p.t.t1;
+	p.t.t1 = p.t.t2;
     return p.t.rdt;
+};
+
+p.t.fps = function() {
+    return (1000 / p.t.dt());
 };
 
 p.t.timer = function(f, fps) {